# src/HomeOps.Contracts/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/family-members": {
      "get": {
        "tags": [
          "Family Members"
        ],
        "operationId": "GetFamilyMembers",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyMemberDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Family Members"
        ],
        "operationId": "CreateFamilyMember",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFamilyMemberRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyMemberDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          }
        }
      }
    },
    "/api/family-members/{memberId}": {
      "get": {
        "tags": [
          "Family Members"
        ],
        "operationId": "GetFamilyMember",
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyMemberDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Family Members"
        ],
        "operationId": "UpdateFamilyMember",
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFamilyMemberRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyMemberDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Family Members"
        ],
        "operationId": "DeleteFamilyMember",
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FamilyMemberDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "displayColor": {
            "type": "string"
          },
          "initials": {
            "type": "string"
          },
          "memberKind": {
            "$ref": "#/components/schemas/FamilyMemberKind"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "avatarV2Config": {
            "$ref": "#/components/schemas/AvatarV2ConfigDto"
          },
          "avatarSelection": {
            "$ref": "#/components/schemas/AvatarSelectionDto"
          }
        }
      },
      "FamilyMemberKind": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Child",
          "Adult"
        ],
        "enum": [
          0,
          1
        ]
      },
      "AvatarV2ConfigDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "headVariant": {
            "type": "string"
          },
          "hairStyle": {
            "type": "string"
          },
          "hairColor": {
            "type": "string"
          },
          "clothingStyle": {
            "type": "string"
          },
          "clothingColor": {
            "type": "string"
          },
          "accessory": {
            "type": "string"
          },
          "accessoryColor": {
            "type": "string"
          }
        }
      },
      "AvatarSelectionDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schemaVersion": {
            "type": "string"
          },
          "selections": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "CreateFamilyMemberRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "memberKind": {
            "$ref": "#/components/schemas/FamilyMemberKind"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "displayColor": {
            "type": "string",
            "nullable": true
          },
          "initials": {
            "type": "string",
            "nullable": true
          },
          "avatarV2Config": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/AvatarV2ConfigDto"
              }
            ]
          },
          "avatarSelection": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/AvatarSelectionDto"
              }
            ]
          }
        }
      },
      "UpdateFamilyMemberRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "displayColor": {
            "type": "string"
          },
          "initials": {
            "type": "string"
          },
          "memberKind": {
            "$ref": "#/components/schemas/FamilyMemberKind"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "avatarV2Config": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/AvatarV2ConfigDto"
              }
            ]
          },
          "avatarSelection": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/AvatarSelectionDto"
              }
            ]
          }
        }
      },
...
    }
  }
}